Require unambiguous listen IP at "swarm init"
There are currently problems with "swarm init" when an explicit
--listen-addr flag is not provided. swarmkit defaults to finding the IP
address associated with the default route, and in cloud setups this is
often the wrong choice.

Instead of trying to use this magic, error out if there is more than one
usable candidate IP address on the system. This requires a user to
explicitly choose instead of letting swarmkit make the wrong choice.

Also support specifying interface names to --listen-addr. This will fail
if the interface has multiple IP addresses (unless it has a single IPv4
address and a single IPv6 address - then we resolve the tie in favor of
IPv4).

Make corresponding API and CLI docs changes.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>