Provide a "local address" for libnetwork to use
For "swarm init", this is the listen address if it's not a wildcard
address, or the advertise address if it's a local address.

For "swarm join", this is the listen address if it's specified
explicitly, otherwise it's resolved by seeing how a connection to the
join address would be routed.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>